dev-requirements.txt
annotated-types==0.7.0
attrs==25.3.0
cattrs==24.1.3
exceptiongroup==1.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
memory-profiler==0.61.0
numpy
orjson
phonenumbers==9.0.4
psutil==7.0.0
pydantic>=1.0.0
pydantic-core>=1.0.0
pygments==2.19.1
python-dateutil==2.9.0.post0
quantities
rich==14.0.0
ruff==0.11.12
six==1.17.0
typing-extensions
urllib3
backports.zoneinfo; python_version < "3.9"
